Version 2.4 Beta 2

+++ osiris/OsirisLib2.01/SampleData.cpp
@@ -151,7 +151,7 @@ const char* SampleData :: GetCapillaryNumber () {
 }
 
 
-int SampleData :: GetDataOffset () {
+long SampleData :: GetDataOffset () {
 
 	return -1;
 }
